Add files via upload
Create a responsive website for the latest movie list using HTML CSS JAVASCRIPT. Display the latest movie list on the page and add details(Name, image, year, rating, language, etc). Options to filter based on rating or language; option to add new movies to the list.  To sort the list based on rating or language etc.